Restructure server.js into modular architecture
